X2APIC, ACPI, related improvements

+++ kernel/arch/aarch64/mod.rs
@@ -249,7 +249,7 @@ impl SmpOps for Aarch64Arch {
         smp::enable_ap_scheduling();
     }
 
-    fn set_bsp_cpu_id(acpi: &mut AcpiInfo, hw_id: u8) {
+    fn set_bsp_cpu_id(acpi: &mut AcpiInfo, hw_id: u32) {
         acpi.bsp_cpu_id = hw_id;
         for cpu in acpi.cpus.iter_mut() {
             cpu.is_bsp = cpu.hw_cpu_id == hw_id;
@@ -269,8 +269,8 @@ impl LocalTimerOps for Aarch64Arch {
         // This function is a no-op on aarch64 since we don't use a LAPIC-style model
     }
 
-    fn current_hw_cpu_id() -> u8 {
-        cpu::cpu_id() as u8
+    fn current_hw_cpu_id() -> u32 {
+        cpu::cpu_id()
     }
 
     fn calibrate_and_start_timer(_vector: u8, interval_ms: u32) -> u32 {
